backend/requirements.txt
amqp==5.2.0
asgiref==3.6.0
astroid==3.1.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.8
Django==4.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.6.1
exceptiongroup==1.2.0
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.6
mccabe==0.7.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathlib2==2.3.7.post1
pathspec==0.5.5
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
Pygments==2.2.0
pylint==3.1.0
pymongo==4.6.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.0
zstandard==0.22.0
